Rename icons to be consistent with images
